requirements.txt
# Dependencias principales
pandas
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
MetaTrader5>=5.0.0
pytz>=2021.1

# Procesamiento de datos y análisis
ta-lib>=0.4.0
pandas-ta>=0.3.0

# Utilidades y sistema
psutil>=5.8.0
python-dateutil>=2.8.1
pywin32>=300; sys_platform == 'win32'

# Logging y configuración
pyyaml>=6.0
colorlog>=6.0.0